backend/requirements.txt
fastapi
uvicorn
google-api-python-client
python-dotenv
importlib-metadata
openai
pytest
httpx

